+++ app/src/main/java/com/example/bettercalc/parsing/Token.kt
@@ -1,6 +1,7 @@
 package com.example.bettercalc.parsing
 
 import com.example.bettercalc.parsing.TokenType
+import java.math.BigDecimal
 
-data class Token(val tokenType: TokenType, val literal: Double?) {
+data class Token(val tokenType: TokenType, val literal: BigDecimal?) {
 }
